more models

+++ Domain/VBMS.Domain/Enums/Gender.cs
@@ -2,7 +2,7 @@
 
 namespace VBMS.Domain.Enums;
 
-public enum Gender : int
+public enum Gender : byte
 {
     None = 0,
     Male,
